test for editor
