+++ package.json
@@ -20,6 +20,7 @@
     "gatsby-plugin-sharp": "^2.6.18",
     "gatsby-remark-copy-linked-files": "^2.3.11",
     "gatsby-remark-images": "^3.3.18",
+    "gatsby-remark-relative-images": "^0.3.0",
     "gatsby-source-filesystem": "^2.3.18",
     "gatsby-theme-codebushi": "1.0.0",
     "gatsby-transformer-json": "^2.4.10",
