+++ src/templates/InternshipTemplate.js
@@ -5,38 +5,39 @@ import Layout from "../components/layout/Layout"
 import Card from "../components/Card"
 import SEO from '../components/SEO'
 import { OutboundLink } from 'gatsby-plugin-google-analytics';
+
 export const query = graphql`
-    query( $slug: String! ) {
-        internshipsJson(slug: { eq: $slug }){
-            title
-            sponsoredBy
-            startDate
-            duration
-            registrationFee
-            registrationLink
-            timings
-            image{
-                childImageSharp{
-                    fluid{
-                        ...GatsbyImageSharpFluid
+    query( $id: String! ) {
+        markdownRemark(id: { eq: $id }){
+            frontmatter {
+                title
+                sponsor
+                start_date
+                duration
+                registration_fee
+                registration_link
+                image {
+                    childImageSharp {
+                    fluid {
+                        ...GatsbyImageSharpFluid_withWebp
+                    }
                     }
                 }
+                contacts {
+                    contact_name
+                    phone_number
+                    linkedin_url
+                }
+                topics_covered
+                key_features
+                work_position
             }
-            contact{
-                name
-                phone
-                linkedin
-            }
-            topics_covered
-            key_features
-            position
-            career_prospects
         }
     }
 `
 
-const InternshipTemplate = ({ data }) => {
-    const internship = data.internshipsJson
+const InternshipTemplate = ({data}) => {
+    const internship = data.markdownRemark.frontmatter
     return (
         <Layout>
             <SEO title={`${internship.title}`} />
@@ -46,21 +47,21 @@ const InternshipTemplate = ({ data }) => {
                         <h1 className="text-4xl lg:text-6xl xl:text-6xl font-raleway font-bold leading-none">
                             {internship.title}
                         </h1>
-                        <p className="mt-6 text-gray-600 font-montserrat">Sponsored by : {internship.sponsoredBy}</p>
+                        <p className="mt-6 text-gray-600 font-montserrat">Sponsored by : {internship.sponsory}</p>
                         <p className="font-montserrat text-3xl lg:text-4xl mt-2 font-light">
-                            {internship.position}
+                            {internship.work_position}
                         </p>
                         <p className="font-montserrat mb-4 text-xl lg:text-2xl mt-2 font-light">
-                            {internship.duration} starting {internship.startDate}<br />{internship.timings}
+                            {internship.duration} <br /> starting {internship.start_date}
                         </p>
                         <p className="font-montserrat mt-4 text-3xl lg:text-4xl font-light">
-                            {internship.registrationFee}
+                            Rs.{internship.registration_fee}
                         </p>
                         <div className="mt-4 md:mt-6 text-xl font-bold font-raleway">
-                            <OutboundLink className="noSelect bg-primary-darker hover:text-primary rounded-full text-center text-white text-xl px-16 py-4" href={internship.registrationLink}>Register</OutboundLink>
+                            <OutboundLink className="noSelect bg-primary-darker hover:text-primary rounded-full text-center text-white text-xl px-16 py-4" href={internship.registration_link}>Register</OutboundLink>
                         </div>
                     </div>
-                    <div className="p-10 mt-12 sm:mt-16 md:mt-6 lg:w-1/2">
+                    <div className="mt-12 sm:mt-16 md:mt-6 lg:w-1/2">
                         <Image fluid={internship.image.childImageSharp.fluid} />
                     </div>
                 </div>
@@ -99,27 +100,27 @@ const InternshipTemplate = ({ data }) => {
                     <h2 className="text-lg md:text-2xl lg:text-3xl font-semibold">Contact and Payment Information</h2>
                     <div className="flex-none flex-wrap items-center justify-center sm:-mx-3 mt-12">
                         {
-                            internship.contact.map((person,i) => (
+                            internship.contacts.map((person, i) => (
                                 <div key={i} className="px-3">
                                     <Card className="mb-8 bg-white">
                                         <div className="inline-block mr-3">
                                             <p className="font-semibold text-xl">
-                                                <OutboundLink className="noSelect text-gray-800" href={person.linkedin}>{person.name}</OutboundLink>
+                                                <OutboundLink className="noSelect text-gray-800" href={person.linkedin_url}>{person.contact_name}</OutboundLink>
                                             </p>
                                         </div>
                                         <div className="w-4 h-4 mr-3 inline-block">
                                             <OutboundLink className="noSelect text-lg text-white" href="https://linkedin.com/company/skillbot">
-                                            <svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-                                                <title>LinkedIn icon</title>
-                                                <path d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433c-1.144 0-2.063-.926-2.063-2.065 0-1.138.92-2.063 2.063-2.063 1.14 0 2.064.925 2.064 2.063 0 1.139-.925 2.065-2.064 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z" fill="teal" />
-                                            </svg>
+                                                <svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                                                    <title>LinkedIn icon</title>
+                                                    <path d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433c-1.144 0-2.063-.926-2.063-2.065 0-1.138.92-2.063 2.063-2.063 1.14 0 2.064.925 2.064 2.063 0 1.139-.925 2.065-2.064 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z" fill="teal" />
+                                                </svg>
                                             </OutboundLink>
                                         </div>
                                         <p className="mt-2 text-md">
                                             Consultant, Skillbot
                                     </p>
                                         <p className="mt-4 text-lg font-bold">
-                                            {person.phone}
+                                            {person.phone_number}
                                         </p>
                                     </Card>
                                 </div>
