+++ static/admin/config.yml
@@ -21,13 +21,46 @@ collections:
     slug: '{{slug}}'
     media_folder: ''
     public_folder: ''
+    path: '{{slug}}/internship'
+    editor:
+      preview : true
+    fields:
+      - { label: 'Template Key', name: 'template', widget: 'select', options: ["internship"] }
+      - { label: 'Registration Status', name: 'status', widget: 'boolean' }
+      - { label: 'Title', name: 'title', widget: 'string' }
+      - { label: 'Position Offered', name: 'work-position', widget: 'string' }
+      - { label: 'Career Prospects', name: 'career-prospects', widget: 'string', required: false }
+      - { label: 'Sponsored By', name: 'sponsor', widget: 'string' }
+      - { label: 'Start Date', name: 'start_date', widget: 'date', format: 'Do MMM YYYY'}
+      - { label: 'Timings', name: 'timings', widget: 'string'}
+      - { label: 'Duration', name: 'duration', widget: 'string' }
+      - { label: 'Registration Fee', name: 'registration_fee', widget: 'number' }
+      - { label: 'Registration Link', name: 'registration_link', widget: 'string' }
+      - { label: 'Image', name: 'image', widget: 'image' }
+      - { label: 'Key Features', name: 'key_features', widget: 'list', field: { label: "Feature", name: "feature", widget: 'string'}}
+      - { label: 'Topics Covered', name: 'topics_covered', widget: 'list', field: { label: "Topic", name: "topic", widget: 'string'}}
+      - label: 'Contacts'
+        name: 'contacts'
+        widget: 'list' 
+        fields: 
+          - { label: "Name", name: "contact_name", widget: 'string'}
+          - { label: "Phone", name: "phone_number", widget: 'string'}
+          - { label: "LinkedIn ID", name: "linkedin_url", widget: 'string'}
+  
+  - name : "skill-dev-program"
+    label : "Skill Development Program"
+    folder : "src/pages/skill-dev-program"
+    create : true
+    slug: '{{slug}}'
+    media_folder: ''
+    public_folder: ''
     path: '{{slug}}/index'
     editor:
       preview : true
     fields:
       - { label: 'Short Title', name: 'slug', widget: 'string' }
+      - { label: 'Status', name: 'status', widget: 'boolean' }
       - { label: 'Title', name: 'title', widget: 'string' }
-      - { label: 'Position Offered', name: 'work-position', widget: 'string' }
       - { label: 'Career Prospects', name: 'career-prospects', widget: 'string', required: false }
       - { label: 'Sponsored By', name: 'sponsor', widget: 'string' }
       - { label: 'Start Date', name: 'start_date', widget: 'date', format: 'Do MMM YY'}
